Add files via upload
purpose: to analysis which stock should be pick per the sales volume and the return 
result: compare 2018 return with the 2017 ones, the overall stock performance was worse in 2018 as the economy was hugely impacted by the President Donald Trump’s trade war with China. As a result, the slowdown in global economic growth and concern that the Federal Reserve was raising interest rates too quickly all contributed to a pessimistic reaction from the stock market. Under this pressure, the overall stock market in 2018 sinked to the bottom. However, RUN has a 78.4% increase . 

There are pro and cons to refactor the original VBA script: 

pros: 
1. you will have chances to add new features or functionality to do more analysis ex. adding a two year comparison 
2. Sometimes bad patterns like tight coupling, duplicate code, long methods, large classes, etc. are detected in the code;  the code should be refactored in this case.
3. codes might be too complex or even wrong somethings. Refactoring could fix those issues. 

cons: 

1. You might accidentally change some good features.