Update README.md
Overall the 2018 stock mark was poorly performed than the 2017. The was due to the President Donald Trump’s trade war with China, the slowdown in global economic growth and concern that the Federal Reserve was raising interest rates too quickly all contributed to a pessimistic reaction from the stock market. However, RUN has a interest of 78.4% in 2018 compared with 
Refactoring could have multiple pros and cons 
Pros: 

Allows you to add new features and detect bugs 
Codes are written badly in some cases, and so many bugs are raised. In this case, fixing of bugs take too much effort. So, the root cause of bugs can be code smell. So, before fixing bugs code should be refactored
Cons: 

Delivery Deadline could be stretched 
The cost of refactoring is higher 
Don't do delayed refactoring because it contains a big mess and makes it very difficult to refactor. Do early refactoring.
Stable code should not be refactored.